+++ inflearn-java-testing/src/test/resources/application-test.properties
@@ -1,7 +1,6 @@
 env = qa
 
-spring.datasource.url=jdbc:postgresql://localhost:15432/studytest
-spring.datasource.username=studytest
-spring.datasource.password=studytest
+spring.datasource.url=jdbc:tc:postgresql:///studytest
+spring.datasource.driver-class-name=org.testcontainers.jdbc.ContainerDatabaseDriver
 
 spring.jpa.hibernate.ddl-auto=create-drop
